+++ package.js
@@ -28,6 +28,7 @@ function configure( api ){
     api.use( 'ecmascript' );
     api.use( 'less@4.0.0', 'client' );
     api.use( 'pwix:i18n@1.3.2' );
+    api.use( 'pwix:modal@1.6.0' );
     api.use( 'tmeasday:check-npm-versions@1.0.2', 'server' );
 }