+++ src/common/js/index.js
@@ -4,5 +4,3 @@
 
 import './global.js';
 import './constants.js';
-//
-import './i18n.js';