+++ src/server/js/check_npms.js
@@ -5,14 +5,11 @@
 import { checkNpmVersions } from 'meteor/tmeasday:check-npm-versions';
 
 if( false ){
-    require( '@popperjs/core/package.json' );
-    require( 'bootstrap/package.json' );
+    // packages white listing
 }
 
 checkNpmVersions({
-    '@popperjs/core': '^2.11.6',
-    'bootstrap': '^5.2.1',
-    'jquery-ui': '^1.13.2'
+    'lodash': '^4.17.21'
 },
     'pwix:bootbox'
 );